requirements.txt
matplotlib==3.10.8
numpy==2.4.4
torch==2.11.0+cu130
torchvision==0.26.0+cu130
tqdm==4.67.3

Pillow

gradio

fastapi
uvicorn
pydantic